Add files via upload
Question:
1. The shortest path from source node to all possible nodes with PRIMS, KRISKALS, DIJKSTRA'S algorithms using C/C++/Python/Java. Give a justification for the choice of the programming language.

2. Give a note of comparison regarding the performance of the algorithms. If you feel that any of the algorithms fail to find any of the required paths, give justification for the same.

Solution:
Prims and Kruskals algorithms can't be used for directed graphs. Whereas, Dijkstras works only for non-negative edges. Therefore neither of the given algorithms are suitable for the given problem. In this case, Djikstras can be used using the help of Bellman Ford algorithm, which specialises in these situation. Hence, I've written a code for Bellman Ford (Djikstras) to find the shortest source path.